Add configurable EvmReferenceResolver
